pyproject.toml
[tool.poetry]
name = "reflection-critique"
version = "0.1.0"
description = ""
authors = ["vasabi07 <vasanthanarulselvan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
langchain = "^0.3.13"
langgraph = "^0.2.60"
langchain-community = "^0.3.13"
loadenv = "^0.1.1"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
